requirements.txt
annotated-types==0.5.0
anyio==3.7.1
Authlib==1.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
mysql==0.0.3
mysqlclient==2.2.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
